pyproject.toml
[tool.poetry]
name = "dongwon-competition"
version = "0.1.0"
description = "A sales simulation tool for Dongwon competition."
authors = ["sunjoonk <sunjoon.dev@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
pandas = "^1.3"
numpy = "^1.21"
python-dotenv = "^0.19"
google-generativeai = "^0.5.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = "src"
testpaths = "tests"